+++ encoding/protobuf/error.go
@@ -27,6 +27,7 @@ import (
 
 	"github.com/gogo/googleapis/google/rpc"
 	"github.com/gogo/protobuf/proto"
+	"github.com/gogo/protobuf/types"
 	"github.com/gogo/status"
 	"go.uber.org/yarpc/api/transport"
 	"go.uber.org/yarpc/internal/grpcerrorcodes"
@@ -45,7 +46,7 @@ var _ error = (*pberror)(nil)
 type pberror struct {
 	code    yarpcerrors.Code
 	message string
-	details []interface{}
+	details []*types.Any
 }
 
 func (err *pberror) Error() string {
@@ -74,7 +75,9 @@ func NewError(code yarpcerrors.Code, message string, options ...ErrorOption) err
 		message: message,
 	}
 	for _, opt := range options {
-		opt.apply(pbErr)
+		if err := opt.apply(pbErr); err != nil {
+			return err
+		}
 	}
 	return pbErr
 }
@@ -93,20 +96,36 @@ func GetErrorDetails(err error) []interface{} {
 	}
 	var target *pberror
 	if errors.As(err, &target) {
-		return target.details
+		results := make([]interface{}, 0, len(target.details))
+		for _, any := range target.details {
+			detail := &types.DynamicAny{}
+			if err := types.UnmarshalAny(any, detail); err != nil {
+				results = append(results, err)
+				continue
+			}
+			results = append(results, detail.Message)
+		}
+		return results
 	}
 	return nil
 }
 
 // ErrorOption is an option for the NewError constructor.
-type ErrorOption struct{ apply func(*pberror) }
+type ErrorOption struct{ apply func(*pberror) error }
 
 // WithErrorDetails adds to the details of the error.
+// If any errors are encountered, it returns the first error encountered.
+// See: https://github.com/gogo/status/blob/master/status.go#L175
 func WithErrorDetails(details ...proto.Message) ErrorOption {
-	return ErrorOption{func(err *pberror) {
+	return ErrorOption{func(err *pberror) error {
 		for _, detail := range details {
-			err.details = append(err.details, detail)
+			any, terr := types.MarshalAny(detail)
+			if terr != nil {
+				return terr
+			}
+			err.details = append(err.details, any)
 		}
+		return nil
 	}}
 }
 
@@ -128,20 +147,14 @@ func convertToYARPCError(encoding transport.Encoding, err error, codec *codec, r
 }
 
 func createStatusWithDetail(pberr *pberror, encoding transport.Encoding, codec *codec) (*yarpcerrors.Status, error) {
-	details := make([]proto.Message, 0, len(pberr.details))
-	for _, detail := range pberr.details {
-		if pbdetail, ok := detail.(proto.Message); ok {
-			details = append(details, pbdetail)
-		} else {
-			return nil, errors.New("proto error detail is not proto.Message compatible")
-		}
+	if pberr.code == yarpcerrors.CodeOK {
+		return nil, errors.New("no status error for error with code OK")
 	}
 
-	st, convertErr := status.New(grpcerrorcodes.YARPCCodeToGRPCCode[pberr.code], pberr.message).WithDetails(details...)
-	if convertErr != nil {
-		return nil, convertErr
-	}
-	detailsBytes, cleanup, marshalErr := marshal(encoding, st.Proto(), codec)
+	st := status.New(grpcerrorcodes.YARPCCodeToGRPCCode[pberr.code], pberr.message).Proto()
+	st.Details = pberr.details
+
+	detailsBytes, cleanup, marshalErr := marshal(encoding, st, codec)
 	if marshalErr != nil {
 		return nil, marshalErr
 	}
@@ -157,15 +170,16 @@ func setApplicationErrorMeta(pberr *pberror, resw transport.ResponseWriter) {
 		return
 	}
 
+	decodedDetails := GetErrorDetails(pberr)
 	var appErrName string
-	if len(pberr.details) > 0 { // only grab the first name since this will be emitted with metrics
+	if len(decodedDetails) > 0 { // only grab the first name since this will be emitted with metrics
 		appErrName = messageNameWithoutPackage(proto.MessageName(
-			pberr.details[0].(proto.Message)),
+			decodedDetails[0].(proto.Message)),
 		)
 	}
 
-	details := make([]string, 0, len(pberr.details))
-	for _, detail := range pberr.details {
+	details := make([]string, 0, len(decodedDetails))
+	for _, detail := range decodedDetails {
 		details = append(details, protobufMessageToString(detail.(proto.Message)))
 	}
 
@@ -208,11 +222,10 @@ func convertFromYARPCError(encoding transport.Encoding, err error, codec *codec)
 		return unmarshalErr
 	}
 
-	details := status.FromProto(st).Details()
-	return newErrorWithDetails(yarpcErr.Code(), yarpcErr.Message(), details)
+	return newErrorWithDetails(yarpcErr.Code(), yarpcErr.Message(), st.GetDetails())
 }
 
-func newErrorWithDetails(code yarpcerrors.Code, message string, details []interface{}) error {
+func newErrorWithDetails(code yarpcerrors.Code, message string, details []*types.Any) error {
 	return &pberror{
 		code:    code,
 		message: message,